account for bug when string isn't in db

+++ deleteCoach.php
@@ -21,12 +21,24 @@
 <?php
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $name = $_SESSION['name'];
-        $query = "DELETE FROM Coaches WHERE name=:name";
-        $statement = $db->prepare($query);
-        $statement->bindValue(':name', $name);
-        $statement->execute();
-        $player_info = $statement->fetchAll();
-        $statement->closeCursor();
+        $query = "SELECT * FROM Players WHERE name =:name";
+        $stmt = $db->prepare($query);
+        $stmt->bindParam(":name", $name);
+        $stmt->execute();
+        $res = $stmt->fetchAll();
+        $found = empty($res) ? 0 : 1;
+        $stmt->closeCursor();
+        if($found){
+            $query = "DELETE FROM Coaches WHERE name=:name";
+            $statement = $db->prepare($query);
+            $statement->bindValue(':name', $name);
+            $statement->execute();
+            $player_info = $statement->fetchAll();
+            $statement->closeCursor();
+        }
+        else{
+            echo "<div class='container' style='text-align: center;'><span class='error_message' id='msg_user'><h4><b>That coach doesn't exists</b></h4></span></div>";
+        }
 }
 ?>
 </html>
